Extract superblock height-based identification
