+++ divi/src/SuperblockHelpers.cpp
@@ -1,0 +1,14 @@
+#include <SuperblockHelpers.h>
+#include <chainparams.h>
+
+bool IsValidLotteryBlockHeight(int nBlockHeight)
+{
+    return nBlockHeight >= Params().GetLotteryBlockStartBlock() &&
+            ((nBlockHeight % Params().GetLotteryBlockCycle()) == 0);
+}
+
+bool IsValidTreasuryBlockHeight(int nBlockHeight)
+{
+    return nBlockHeight >= Params().GetTreasuryPaymentsStartBlock() &&
+            ((nBlockHeight % Params().GetTreasuryPaymentsCycle()) == 0);
+}
