+++ divi/src/SuperblockHelpers.h
@@ -1,0 +1,5 @@
+#ifndef SUPERBLOCK_HELPERS_H
+#define SUPERBLOCK_HELPERS_H
+bool IsValidLotteryBlockHeight(int nBlockHeight);
+bool IsValidTreasuryBlockHeight(int nBlockHeight);
+#endif // SUPERBLOCK_HELPERS_H
